third_party/blink/web_tests/platform/generic/html/document_metadata/base-multiple-expected.txt

Test the behavior of multiple base elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS clean(anchor.href) is 'http://originalbase.com/file'
PASS document.head.appendChild(base), clean(anchor.href) is 'http://domain.com/base/file'
PASS base.href = 'http://domain.com/base-changed/', clean(anchor.href) is 'http://domain.com/base-changed/file'
PASS document.head.removeChild(base), clean(anchor.href) is 'http://originalbase.com/file'
PASS document.head.appendChild(base), document.head.appendChild(base2), clean(anchor.href) is 'http://domain.com/base/file'
PASS base.removeAttribute('href'), clean(anchor.href) is 'http://domain.com/base2/file'
PASS document.head.appendChild(base3), clean(anchor.href) is 'http://domain.com/base2/file'
PASS successfullyParsed is true

TEST COMPLETE
